==> script/01_Data_Extraction/p_01_UKBB_CSV.ipynb <==
### Packages

# pip install webdriver-manager

# pip install selenium
# pip install webdriver-manager

# Packages for webscraping
from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from webdriver_manager.chrome import ChromeDriverManager
from selenium.common.exceptions import NoSuchElementException
from selenium.webdriver.common.by import By

#Package to handle dataframes
import pandas as pd

#Package for progress bar
from tqdm import tqdm

#### Configure webdriver to use Chrome browser

s=Service(ChromeDriverManager().install())
driver = webdriver.Chrome(service=s)

#### ID selection (phenotypes)

### Add
# 6177, 6153, 1588, 1578, 1608, 5364, 1568, 1598, 904, 6138, 20116, 1289, 1309, 2040, 20023, 1200, 1498



# Select UKBB ids of interest
Biomarker_CAL = [30600,30610,30620,30630,30640,30650,30840,2744,20022,3148,21001,23105,23100,30680,30690,30700,
              30710,30720,4079,30150,20016,3062,30730,30740,46,47,30750,30760,102,50,30770,30780,30790,30120,
              30050,30130,20127,30140,30810,30080,30010,30250,30830,4080,30850,30870,30860,30880,30670,30890,
              30000,21002,48,49, #Attention 46,47,48,49 ratio in Auwerx Phenotypes (57 Phenos, some moved)
                 22192, 30800, 22426] # Remove 51, 22433

# Need to calculate and mean grip strength  WHR, WHR adjusted for BMI
# 2714, 3581, 2395 and 2375 moved to Life history

print("Nb of biomarkers:", len(Biomarker_CAL))

Disease_NTL = []

print("Nb of diseases:", len(Disease_NTL))

Lifestyle_SML = [1160, 20117, 1558, 100940, 1369, 1408, 1528, 1110, 2237, 2644, 189, 845,
                738, 2267] # 20118 might wanna remove Urban/Rural

#Remove 21067, 1031, 20447, 20433

print("Nb of lifstyle phenotypes:", len(Lifestyle_SML))

Life_history_SML = [40007, 1807, 3526, 2754, 2764, 2714, 3581, 2395, 2375, 40008, 21003, 2217, 2207, 2734]

#Remove 3393

# Difference between 2714 and 3581 should be calculated (female repoductive period)

print("Nb of life history phenotypes:", len(Life_history_SML))

print("Total:", len(Biomarker_CAL) + len(Disease_NTL) + len(Lifestyle_SML) + len (Life_history_SML))
print("Check no recurrency:", len(set().union(Biomarker_CAL,Disease_NTL,Lifestyle_SML,Life_history_SML)))

### Webscraping function

def UKBB_to_DF(ukbbidlist):
    # Initiate empty lists to store data of interest
    Description=[]
    Category = []
    ValueType=[]
    Nbpart = []
    Sex=[] 
    Mean=[]
    Median=[]
    Stdv=[]

    # 
    for ukbbid in tqdm(ukbbidlist):
        driver = webdriver.Chrome(service=s)
        driver.get("https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=" + str(ukbbid))
    # Search by XPATH (put XPATH between """___""")
        Description.append(driver.find_element(by=By.XPATH, value="""//*[@id="main"]/table[1]/tbody/tr[1]/td[2]""").text)
        Category.append(driver.find_element(by=By.XPATH, value="""//*[@id="main"]/table[1]/tbody/tr[2]/td[2]/a[1]""").text)
        ValueType.append(driver.find_element(by=By.XPATH, value="""//*[@id="main"]/table[2]/tbody/tr[1]/td[5]""").text)
        Nbpart.append(int(driver.find_element(by=By.XPATH, value="""//*[@id="main"]/table[2]/tbody/tr[1]/td[2]""").text.replace(",", "")))
        Sex.append(driver.find_element(by=By.XPATH, value="""//*[@id="main"]/table[2]/tbody/tr[1]/td[8]""").text)
    # Search by text element (write search query e.g. "Mean = " and remove by replacing string)
        try:
            Mean.append(driver.find_element(By.XPATH, value = ("//*[contains(text(),'Mean = ')]")).text.replace("Mean = ", ""))
        except NoSuchElementException:
            Mean.append("NA")
    
        try:    
            Median.append(driver.find_element(By.XPATH, value = ("//*[contains(text(),'Median')]/following::td")).text)
        except NoSuchElementException:
            Median.append("NA")

        try:
            Stdv.append(driver.find_element(By.XPATH, value = ("//*[contains(text(),'Std.dev = ')]")).text.replace("Std.dev = ", ""))
        except NoSuchElementException:
            Stdv.append("NA")

        driver.close()
    # Write dataframe
    df = pd.DataFrame({'FieldID':ukbbidlist,'Description':Description, "Category":Category,
                       "Value_Type":ValueType, "Nb_Participants": Nbpart,
                       'Sex':Sex,"Mean":Mean, "Median": Median, "SD":Stdv}) 
    
    return(df)

### Get data as dataframe

df1 = UKBB_to_DF(Biomarker_CAL)
df2 = UKBB_to_DF(Disease_NTL)
df3 = UKBB_to_DF(Lifestyle_SML)
df4 = UKBB_to_DF(Life_history_SML)

frames = [df1, df2, df3, df4]
df = pd.concat(frames)
df.head(10)

df.to_csv('../', index=False, encoding='utf-8')
df.to_excel('../', index=False, encoding='utf-8')

### From TL_metadata file

### Load dataframe
df_metadata = pd.read_csv(r'../TL_metadata.csv')
### Get list of FieldIDs
fieldids = list(df_metadata["FieldID"])
fieldids = [x for x in fieldids if str(x) != 'nan'] # Remove NaN
fieldids = [int(i) for i in fieldids] # Convert float to int

### Extract information
df = UKBB_to_DF(fieldids)

df

df.to_csv('../', index=False, encoding='utf-8')
df.to_excel('../', index=False, encoding='utf-8')
